Add head info in each page instead of routes.js
