+++ src/pages/Landing.js
@@ -9,7 +9,8 @@ import {
   buildQuestionsFromLangObj,
   FullWidth,
   FadeIn,
-  FilledButton
+  FilledButton,
+  PageHead
 } from 'components/Marketing';
 import mixpanel from 'mixpanel-browser';
 import { Routes } from 'utils/constants';
@@ -508,6 +509,11 @@ function Landing() {
 
   return (
     <MarketingLayout>
+      <PageHead
+        title={lang.landing_page.meta.title}
+        description={lang.landing_page.meta.description}
+        imgUrl="https://oasis.app/meta/Oasis-app.png"
+      />
       <Content>
         <Box mt={{ s: '126px', m: '149px' }} px={{ s: '10px', m: 0 }}>
           <Text.h1>{lang.landing_page.headline}</Text.h1>
